+++ src/raws/lua_bridge.hpp
@@ -32,8 +32,8 @@ void read_lua_table(const std::string &table, const std::function<void(std::stri
 void read_lua_table_inner(const std::string &table, const std::function<void(std::string)> &functor);
 void read_lua_table_inner_p(const std::string &table, const std::function<void(std::string)> &on_start, const std::function<void(std::string)> &on_end, const lua_parser &parser);
 inline std::string lua_str() { return lua_tostring(lua_state, -1); }
-inline int lua_int() { return lua_tonumber(lua_state, -1); }
-inline float lua_float() { return lua_tonumber(lua_state, -1); }
+inline int lua_int() { return static_cast<int>(lua_tonumber(lua_state, -1)); }
+inline float lua_float() { return static_cast<float>(lua_tonumber(lua_state, -1)); }
 bengine::color_t read_lua_color(std::string field);
 
 template<class T>
