+++ src/raws/profession.cpp
@@ -9,7 +9,7 @@ std::vector<profession_t> starting_professions;
 
 profession_t * get_random_profession(bengine::random_number_generator &rng) {
 	const auto number_of_professions = starting_professions.size();
-	const auto selected_profession = rng.roll_dice(1, number_of_professions -1);
+	const auto selected_profession = rng.roll_dice(1, static_cast<int>(number_of_professions) -1);
 	return &starting_professions[selected_profession];
 }
